refactor(api): extract client-AI exchange resolver for reuse by the neutral office-addin exchange
