+++ apps/api/src/routes/clientAi/auth.ts
@@ -1,11 +1,5 @@
 import { Hono } from 'hono';
 import { zValidator } from '../../lib/validation';
-import { and, eq } from 'drizzle-orm';
-import { nanoid } from 'nanoid';
-import { db, withSystemDbAccessContext } from '../../db';
-import { portalUsers } from '../../db/schema';
-import { clientAiTenantMappings } from '../../db/schema/clientAi';
-import { organizations, partners } from '../../db/schema/orgs';
 import { getRedis } from '../../services/redis';
 import { rateLimiter } from '../../services/rate-limit';
 import { getTrustedClientIp, rateLimitIpKey } from '../../services/clientIp';
@@ -16,59 +10,18 @@ import {
   ClientAiEntraInvalidTokenError,
   ClientAiEntraJwksUnavailableError,
 } from '../../services/clientAiEntraJwt';
-import { getOrgPolicy, isClientUserPermitted } from '../../services/clientAiPolicy';
-import {
-  exchangeSchema,
-  CLIENT_AI_REDIS_KEYS,
-  CLIENT_AI_SESSION_TTL_SECONDS,
-  EXCHANGE_RATE_LIMIT,
-} from './schemas';
+import { resolveAndMintClientSession } from '../../services/clientAiExchange';
+import { exchangeSchema, EXCHANGE_RATE_LIMIT } from './schemas';
 
 /**
  * POST /client-ai/auth/exchange — Entra ID token → Breeze client-AI session.
  * Spec §3. Pre-auth route: tenant context comes FROM the verified token (tid →
- * client_ai_tenant_mappings), so DB work runs under system scope. One fast DB
- * block; Redis work stays outside it (#1105).
+ * client_ai_tenant_mappings), so DB work runs under system scope. Resolution
+ * (tenant mapping, policy checks, JIT provisioning, session mint) lives in
+ * `services/clientAiExchange.ts` — this handler is just gates + audit + wire shape.
  */
 
 export const clientAiAuthRoutes = new Hono();
-
-type ExchangeUser = {
-  id: string;
-  orgId: string;
-  email: string;
-  name: string | null;
-  status: string;
-};
-
-type Denied = {
-  denied: {
-    status: 403 | 404;
-    error: string;
-    orgId: string | null;
-    details: Record<string, unknown>;
-  };
-};
-/** White-label footer fields (spec §11), sourced from the org policy's branding JSONB. */
-type ExchangeBranding = { displayName: string | null; logoUrl: string | null };
-type Resolved = { user: ExchangeUser; provisioned: boolean; branding: ExchangeBranding };
-
-/** policy.branding is free-form JSONB — pull only the two known string fields, coercing anything else to null. */
-function brandingFromPolicy(branding: Record<string, unknown> | null | undefined): ExchangeBranding {
-  const asString = (v: unknown): string | null => (typeof v === 'string' ? v : null);
-  return {
-    displayName: asString(branding?.displayName),
-    logoUrl: asString(branding?.logoUrl),
-  };
-}
-
-const USER_COLUMNS = {
-  id: portalUsers.id,
-  orgId: portalUsers.orgId,
-  email: portalUsers.email,
-  name: portalUsers.name,
-  status: portalUsers.status,
-};
 
 function auditExchange(
   c: RequestLike,
@@ -131,171 +84,24 @@ clientAiAuthRoutes.post('/auth/exchange', zValidator('json', exchangeSchema), as
     throw err;
   }
 
-  const resolution = await withSystemDbAccessContext(async (): Promise<Denied | Resolved> => {
-    const [mapping] = await db
-      .select({
-        orgId: clientAiTenantMappings.orgId,
-        partnerEnabled: partners.aiForOfficeEnabled,
-      })
-      .from(clientAiTenantMappings)
-      .innerJoin(organizations, eq(organizations.id, clientAiTenantMappings.orgId))
-      .innerJoin(partners, eq(partners.id, organizations.partnerId))
-      .where(eq(clientAiTenantMappings.entraTenantId, claims.tid))
-      .limit(1);
+  const outcome = await resolveAndMintClientSession(claims, redis);
 
-    if (!mapping) {
-      return {
-        denied: {
-          status: 404,
-          error: 'tenant_not_provisioned',
-          orgId: null,
-          details: { reason: 'tenant_not_provisioned', tid: claims.tid },
-        },
-      };
-    }
-
-    // Per-partner entitlement gate (the cost gate): no enabled partner ⇒ no
-    // session ⇒ no AI spend. Sits above the per-org policy.enabled check below.
-    if (!mapping.partnerEnabled) {
-      return {
-        denied: {
-          status: 403,
-          error: 'disabled',
-          orgId: mapping.orgId,
-          details: { reason: 'partner_not_enabled', tid: claims.tid, oid: claims.oid },
-        },
-      };
-    }
-
-    const policy = await getOrgPolicy(mapping.orgId);
-    if (!policy.enabled) {
-      return {
-        denied: {
-          status: 403,
-          error: 'disabled',
-          orgId: mapping.orgId,
-          details: { reason: 'disabled', tid: claims.tid, oid: claims.oid },
-        },
-      };
-    }
-
-    const now = new Date();
-    let provisioned = false;
-    let [user] = await db
-      .select(USER_COLUMNS)
-      .from(portalUsers)
-      .where(
-        and(eq(portalUsers.entraTenantId, claims.tid), eq(portalUsers.entraOid, claims.oid))
-      )
-      .limit(1);
-
-    if (!user) {
-      // portal_users.email is NOT NULL; some Entra token shapes carry no usable
-      // address — fall back to a synthetic, non-routable one.
-      const email = claims.email ?? `${claims.oid}@${claims.tid}.entra.invalid`;
-      try {
-        const inserted = await db
-          .insert(portalUsers)
-          .values({
-            orgId: mapping.orgId,
-            email,
-            name: claims.name,
-            passwordHash: null,
-            entraOid: claims.oid,
-            entraTenantId: claims.tid,
-            authMethod: 'entra',
-            lastLoginAt: now,
-          })
-          .returning(USER_COLUMNS);
-        user = inserted[0];
-        provisioned = true;
-      } catch (err) {
-        // Concurrent first-exchange race: portal_users_entra_identity_uniq
-        // makes the loser 23505 — re-select the winner's row.
-        if ((err as { cause?: { code?: string } }).cause?.code !== '23505') throw err;
-        [user] = await db
-          .select(USER_COLUMNS)
-          .from(portalUsers)
-          .where(
-            and(eq(portalUsers.entraTenantId, claims.tid), eq(portalUsers.entraOid, claims.oid))
-          )
-          .limit(1);
-      }
-    } else {
-      await db
-        .update(portalUsers)
-        .set({ lastLoginAt: now, updatedAt: now, ...(claims.name ? { name: claims.name } : {}) })
-        .where(eq(portalUsers.id, user.id));
-    }
-
-    if (!user) {
-      return {
-        denied: {
-          status: 403,
-          error: 'provisioning_failed',
-          orgId: mapping.orgId,
-          details: { reason: 'provisioning_failed', tid: claims.tid, oid: claims.oid },
-        },
-      };
-    }
-
-    if (user.status !== 'active') {
-      return {
-        denied: {
-          status: 403,
-          error: 'account_inactive',
-          orgId: mapping.orgId,
-          details: { reason: 'account_inactive', portalUserId: user.id },
-        },
-      };
-    }
-
-    if (!isClientUserPermitted(policy, user.id)) {
-      return {
-        denied: {
-          status: 403,
-          error: 'user_not_permitted',
-          orgId: mapping.orgId,
-          details: { reason: 'user_not_permitted', portalUserId: user.id },
-        },
-      };
-    }
-
-    return { user, provisioned, branding: brandingFromPolicy(policy.branding) };
-  });
-
-  if ('denied' in resolution) {
+  if (outcome.kind === 'denied') {
     auditExchange(c, {
-      orgId: resolution.denied.orgId,
+      orgId: outcome.audit.orgId,
       result: 'denied',
-      details: resolution.denied.details,
+      details: outcome.audit.details,
     });
-    return c.json({ error: resolution.denied.error }, resolution.denied.status);
+    return c.json(outcome.body, outcome.status);
   }
 
-  const { user, provisioned, branding } = resolution;
-  const token = nanoid(48);
-  await redis.setex(
-    CLIENT_AI_REDIS_KEYS.session(token),
-    CLIENT_AI_SESSION_TTL_SECONDS,
-    JSON.stringify({ portalUserId: user.id, orgId: user.orgId, createdAt: new Date().toISOString() })
-  );
-  await redis.sadd(CLIENT_AI_REDIS_KEYS.userSessions(user.id), token);
-  await redis.expire(CLIENT_AI_REDIS_KEYS.userSessions(user.id), CLIENT_AI_SESSION_TTL_SECONDS * 2);
-
   auditExchange(c, {
-    orgId: user.orgId,
+    orgId: outcome.audit.orgId,
     result: 'success',
-    actorId: user.id,
-    actorEmail: user.email,
-    details: { tid: claims.tid, oid: claims.oid, provisioned },
+    actorId: outcome.audit.actorId,
+    actorEmail: outcome.audit.actorEmail,
+    details: outcome.audit.details,
   });
 
-  return c.json({
-    accessToken: token,
-    expiresInSeconds: CLIENT_AI_SESSION_TTL_SECONDS,
-    user: { id: user.id, email: user.email, name: user.name },
-    org: { id: user.orgId },
-    branding,
-  });
+  return c.json(outcome.body);
 });
